userspace: split manager_ha.go by responsibility
pkg/dataplane/userspace/manager_ha.go had reached 2,299 LOC and sat in
the [REFACTOR] tier of the modularity audit. Its size was a symptom: the
HA-shaped filename covered seven unrelated responsibilities, so anyone
touching the session mirror had to read the HA state machine and vice
versa, and the file's lock-discipline invariants (#5007 one-snapshot
build, #5698 pair-transmit contiguity, #5380 transport fast-fail) were
interleaved with counter plumbing that has no locking contract at all.

Split by responsibility, not by line count. What each file owns:

  manager_ha.go (873) — the HA state machine: redundancy-group
    ownership (UpdateRGActive), the watchdog heartbeat and its
    update_ha_state IPC throttle, the helper HA-state publish/clear
    including the #5487 retry debt, HA group inventory seeding and shim
    map refresh, the desired forwarding-arm reconcile, and takeover
    readiness.
  manager_sessions.go (569) — session-table mutation entry points: the
    dataplane.DataPlane install/delete/batch/clear verbs, the #5305
    transactional cluster-synced install, the chunked authoritative
    helper deletes, and the bulk event-stream session export.
  manager_sessionsync_request.go (327) — session-sync request
    construction: BPF key/value to SessionSyncRequest, egress/owner-RG/
    tunnel-endpoint resolution against m.lastSnapshot, zone-name lookup
    and the address/port/MAC encoders. Pure snapshot reads, no I/O.
  manager_counters.go (267) — counter bridging: aggregate the helper's
    per-binding counters from a status poll and push the deltas into the
    BPF global-counter, NAT-rule and per-zone offset maps.
  manager_sessionsync_transmit.go (187) — session-sync transmit, and the
    single owner of its lock discipline: the m.mu drop for socket I/O,
    per-request m.sessionMu on the interleavable bulk path, one hold for
    the whole group on the contiguous pair path, transport fast-fail.
  manager_fabric_sync.go (102) — fabric-state publish to the helper plus
    the #5306 resolved-fabric writeback into m.lastSnapshot.
  napi_probe_interfaces.go (57) — the deterministic Linux netdev list
    the mlx5 NAPI bootstrap probe sends traffic on.

The split forces no new exported surface: every identifier keeps its
name, receiver and visibility, and the package is unchanged, so no test
needed editing to keep compiling.

PURE CODE MOTION. Each declaration was copied by exact source line range
by a script, never retyped, so no body, signature, receiver, visibility,
struct field order, lock scope or defer ordering changed. Equivalence
was proven mechanically by two INDEPENDENT parsers rather than by
reading the diff:

  1. A block parser split the pre-split file and all seven post-split
     files into (leading comment run + declaration) blocks and compared
     them by name: 80 of 80 declarations byte-identical INCLUDING
     indentation, 0 dropped, 0 added. Each declaration's doc comment was
     compared as its own string, so a comment that landed on the wrong
     neighbour fails the check rather than hiding inside a whole-block
     compare.
  2. A go/parser + go/printer dump keyed by declaration name emitted
     each decl's ast.Doc text and printed source; the before and after
     dumps diff to zero over the same 80 declarations. This proof does
     not share code with the splitter or with proof 1.

Six doc and Rust-comment pointers named manager_ha.go as the home of a
symbol that moved; they now name the new file. The refactor audit was
regenerated: manager_ha.go leaves the [REFACTOR] tier and no new file
enters the >=1500 LOC audit, so the change is one deleted row.

Validation (exit codes read directly, never through a pipe): go build
./... rc=0; go vet ./pkg/dataplane/... rc=0; go test -count=1
./pkg/dataplane/... rc=0; go test -count=1 -race -run 'Session|HA'
./pkg/dataplane/userspace/... rc=0; go test -count=1
./pkg/refactoraudit/... rc=0; go test -count=1 ./... rc=0. The _5007,
_5698, _5305, _5380 and _5881 regression cells pass unchanged.

This is HA and session-sync code, so make test-failover is OWED. It was
deliberately NOT run here and the shared loss cluster was not touched.

Refs #6427
