+++ docs/refactoring-audit-current.txt
@@ -11,7 +11,6 @@
 [REFACTOR]   2375  userspace-dp/src/afxdp/worker/loop_body/mod.rs
 [REFACTOR]   2358  pkg/daemon/daemon_nft.go
 [REFACTOR]   2310  userspace-dp/src/afxdp/neighbor.rs
-[REFACTOR]   2299  pkg/dataplane/userspace/manager_ha.go
 [REFACTOR]   2288  pkg/ddns/surface_a.go
 [REFACTOR]   2258  userspace-dp/src/afxdp/cos/queue_service/mod.rs
 [REFACTOR]   2227  userspace-dp/src/session/mod.rs
